clarifying comments
updated the comment on RandomObject to be more clear, but if you are
reading this you it may not still be clear. If we have only one obj
of a particular type, the logic above will cause it not to be found.

this is an nbd thing, because this method will probably die/be unused.